Rename: 인터페이스 I 접두어 제거, 클래스 Impl 접미사 추가
