+++ app/src/main/java/kr/co/wdtt/nbdream/data/repository/WeatherForecastRepositoryImpl.kt
@@ -4,15 +4,14 @@ import kotlinx.coroutines.flow.Flow
 import kr.co.wdtt.nbdream.BuildConfig
 import kr.co.wdtt.nbdream.data.mapper.EntityWrapper
 import kr.co.wdtt.nbdream.data.mapper.WeatherForecastMapper
-import kr.co.wdtt.nbdream.data.remote.api.NetworkApi
 import kr.co.wdtt.nbdream.data.remote.dto.WeatherForecastResponse
-import kr.co.wdtt.nbdream.data.remote.retrofit.INetworkFactoryManager
+import kr.co.wdtt.nbdream.data.remote.retrofit.NetworkFactoryManager
 import kr.co.wdtt.nbdream.domain.entity.WeatherForecastEntity
 import kr.co.wdtt.nbdream.domain.repository.WeatherForecastRepository
 import javax.inject.Inject
 
 internal class WeatherForecastRepositoryImpl @Inject constructor(
-    network: INetworkFactoryManager,
+    network: NetworkFactoryManager,
     private val mapper: WeatherForecastMapper,
 ) : WeatherForecastRepository {
